+++ modules/plugin/twitter.js
@@ -5,7 +5,8 @@ const exec = promisify(require('child_process').exec);
 const HttpsProxyAgent = require("https-proxy-agent");
 const fs = require('fs-extra');
 
-const PROXY_CONF = require("../../config.json").proxy;
+const CONFIG = require("../../config.json");
+const PROXY = CONFIG.proxy;
 const DB_PORT = 27017;
 const DB_PATH = "mongodb://127.0.0.1:" + DB_PORT;
 const BEARER_TOKEN = "Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA";
@@ -43,26 +44,30 @@ function twitterReply(replyMsg) {
 }
 
 /** 检查网络情况，如果连不上Twitter那后面都不用做了*/
-function checkConnection() {
+async function checkConnection() {
     return axios.get("https://twitter.com", {
         headers : {
             "User-Agent" : "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/87.0.4280.88 Safari/537.36"
         }
-    }).then(res => {connection = (res.status == 200) ? true : false})
-    .catch(err => {
-        console.error("Twitter checkConnection error with", err.response.status, err.response.statusText);
+    }).then(res => {
+        if (res.status == 200) {
+            connection = true;
+            console.log("Twitter successfully connected");
+        }
+        return true;
+    }).catch(err => {
+        console.error("Twitter checkConnection error with ",
+            err.response ? [err.response.status, err.response.statusText].join(" ") 
+            : [err.errno, err.code].join(" "));
         return false;
     });
 }
 
 function setAgent() {
-    if (PROXY_CONF.host.startsWith("http") && PROXY_CONF.port != 0) {
+    if (PROXY.startsWith("http")) {
         axios = Axios.create({
-            httpsAgent : new HttpsProxyAgent({
-                hostname : PROXY_CONF.host,
-                port : PROXY_CONF.port,
-                protocol : /^https/.test(PROXY_CONF.host) ? "https" : "http"
-            })
+            proxy: false,
+            httpsAgent : new HttpsProxyAgent(PROXY)
         });
     }
     else axios = Axios;
@@ -92,8 +97,8 @@ function toOptNl(option) {
 }
 
 function firstConnect() {
-    checkConnection().then(() => {
-        if (!connection) {
+    checkConnection().then(res => {
+        if (!res) {
             console.error("Twitter无法连接，功能暂停");
         }
         else {
@@ -107,15 +112,8 @@ function firstConnect() {
                 setTimeout(getCookie, 1000);
             }, 1*60*60*1000);
         }
-    });
-}
-
-function sizeCheck(url) {
-    return axios.get(url).then(res => {
-        return parseInt(res.headers["content-length"]) < MAX_SIZE ? true : false;
     }).catch(err => {
-        console.error(url, err.response.status);
-        return false;
+        console.error("Twitter无法连接，功能暂停");
     });
 }
 
@@ -151,7 +149,12 @@ function getGuestToken() {
         method : "POST",
         url : "https://api.twitter.com/1.1/guest/activate.json",
         headers : headers
-    }).then(res => {guest_token = res.data.guest_token;}).catch(err => console.log(err.response.data))
+    }).then(res => {guest_token = res.data.guest_token;
+    }).catch(err => {
+        console.error("Twitter getGuestToken error with ", 
+        err.response ? [err.response.status, err.response.statusText].join(" ") 
+        : [err.errno, err.code].join(" "));
+    });
 }
 
 /** 获取一个cookie，后面要用*/
@@ -177,7 +180,11 @@ function getCookie() {
             else if (temp = /(_twitter_sess=.+?);/.exec(res.headers["set-cookie"][i])) twitter_sess = temp[1];
         }
         cookie = `dnt=1; fm=0; csrf_same_site_set=1; csrf_same_site=1; gt=${guest_token}; ${ct0}${guest_id}${personalization_id}${twitter_sess}`;
-    }).catch(err => console.error('Twitter cookie设置出错，错误：', err.response.status, err.response.statusText));
+    }).catch(err => {
+        console.error('Twitter getCookie error ', 
+        err.response ? [err.response.status, err.response.statusText].join(" ") 
+        : [err.errno, err.code].join(" "));
+    });
 }
 
 /** 
@@ -186,7 +193,7 @@ function getCookie() {
  * @param {string} tweet_id_str 单条Tweet id
  * @returns {Promise} Tweet Object，如果错误，结果为false
  */
-function getSingleTweet(tweet_id_str) {
+async function getSingleTweet(tweet_id_str) {
     return axios({
         method:'GET',
         url: "https://api.twitter.com/1.1/statuses/show.json",
@@ -207,7 +214,9 @@ function getSingleTweet(tweet_id_str) {
         }
     }).then(res => {return res.data; 
     }).catch(err => {
-        console.error(err.response.data);
+        console.error('Twitter getSingleTweet error with ', 
+            err.response ? [err.response.status, err.response.statusText].join(" ") 
+            : [err.errno, err.code].join(" "));
         return false;
     });
 }
@@ -219,23 +228,43 @@ function getSingleTweet(tweet_id_str) {
  * @param {number} count 获取数量，最大为200
  * @returns {Promise} 用户时间线，如果错误结果为false
  */
-function getUserTimeline(user_id, count = 2, include_rt = false, exclude_rp = true, since_id = "1") {
+async function getUserTimeline(user_id, count = 20, include_rt = 0, include_rp = 0) {
     return axios({
         method:'GET',
-        url: "https://api.twitter.com/1.1/statuses/user_timeline.json",
-        headers : {"Authorization" : BEARER_TOKEN},
+        url: `https://twitter.com/i/api/2/timeline/profile/${user_id}.json`,
+        headers : httpHeader(),
         params : {
             // screen_name : screen_name,
-            "user_id" : user_id,
+            "userId" : user_id,
             "count" : count,
-            "exclude_replies" : exclude_rp,
-            "include_rts" : include_rt,
+            "include_tweet_replies" : include_rp,
+            "include_want_retweets" : include_rt,
             "tweet_mode" : "extended",
-            since_id : since_id
+            "include_cards" : "1",
+            "cards_platform" : "Web-12",
+            "include_ext_alt_text": "true",
+            "include_ext_media_color" : "true",
+            "include_ext_media_availability" : "true",
+            "include_entities" : "true",
+            "include_ext_alt_text" : "true",
+            "include_card_uri" : "true"
         }
-    }).then(res => {return res.data;
+    }).then(res => {
+        let tweets = [];
+        let user = res.data.globalObjects.users[user_id];
+        for (let tweetid of Object.keys(res.data.globalObjects.tweets)) {
+            let tweet = res.data.globalObjects.tweets[tweetid];
+            tweet.user = {name: user.name, screen_name: user.screen_name};
+
+            tweets.push(tweet);
+        }
+        
+        tweets = tweets.sort((a, b) => {return (a.id_str > b.id_str) ? -1 : 1;});
+        return tweets;
     }).catch(err => {
-        console.error(err.response.data);
+        console.error('Twitter getUserTimeline error with ', 
+            err.response ? [err.response.status, err.response.statusText].join(" ") 
+            : [err.errno, err.code].join(" "));
         return false;
     });
 }
@@ -248,7 +277,7 @@ function getUserTimeline(user_id, count = 2, include_rt = false, exclude_rp = tr
  * @param {string} name 用户名称
  * @returns {Promise} user_object，如果没有或者错误会返回false
  */
-function searchUser(name) {
+async function searchUser(name) {
     let header = httpHeader();
     header["x-guest-token"] = guest_token;
     return axios({
@@ -261,7 +290,9 @@ function searchUser(name) {
         }
     }).then(res => {return res.data[0]
     }).catch(err => {
-        console.error("Twitter searchUser error\n" + err.response.data);
+        console.error('Twitter searchUser error with ', 
+            err.response ? [err.response.status, err.response.statusText].join(" ") 
+            : [err.errno, err.code].join(" "));
         return false;
     })
 }
@@ -361,21 +392,25 @@ function unSubscribe(name, context) {
 /**
  * 每过x分钟检查一次订阅列表，如果订阅一个Twitter账号的群的数量是0就删除
  */
-function checkTwiTimeline() {
+async function checkTwiTimeline() {
     if (!connection) return;
-    let check_interval = 7 * 60 * 1000;
-
-    setInterval(async () => {
+    const mongo = await mongodb(DB_PATH, {useUnifiedTopology: true}).connect();
+    const twitter_db = mongo.db('bot').collection('twitter');
+    let subscribes = await twitter_db.find({}).toArray();
+    let check_interval = subscribes.length > 0 ? subscribes.length * 30 * 1000 : 5 * 60 * 1000;
+    mongo.close();
+    
+    async function refreshTimeline() {
         await mongodb(DB_PATH, {useUnifiedTopology: true}).connect().then(async mongo => {
             const twitter_db = mongo.db('bot').collection('twitter');
             const group_option = mongo.db('bot').collection('group_option');
-            const twe_sum = mongo.db('bot').collection('twe_sum');
             let subscribes = await twitter_db.find({}).toArray();
             let options = await group_option.find({}).toArray();
-            let summ = await twe_sum.find({}, {$projection : {list : 0}}).toArray();
 
             if (subscribes.length > 0 && options.length > 0) {
                 i = 0;
+                check_interval = subscribes.length * 30 * 1000;
+                setTimeout(refreshTimeline, check_interval + 30000);
                 checkEach();
             }
             else if (subscribes.length < 1 || options.length < 1) {
@@ -391,34 +426,24 @@ function checkTwiTimeline() {
                 setTimeout(async() => {
                     process: try {
                         if (subscribes[i] == undefined) break process;
-                        let tweet_list = await getUserTimeline(subscribes[i].uid, 5, true, false, subscribes[i].tweet_id);
-                        if (tweet_list != undefined && tweet_list.length > 0 && tweet_list[0].id_str > subscribes[i].tweet_id) {
-                            let groups = subscribes[i].groups;
-                            let bbq_group = [];
+                        let curr_s = subscribes[i];
+                        let tweet_list = await getUserTimeline(curr_s.uid, 5, 1, 1);
+                        if (tweet_list != undefined && tweet_list.length > 0 && tweet_list[0].id_str > curr_s.tweet_id) {
+                            tweet_list = tweet_list.filter(t => t.id_str > curr_s.tweet_id);
+                            let groups = curr_s.groups;
                             let url_list = [];
                             for (let group_id of groups) {
                                 let option = false;
                                 let post = false;
-                                let count = false;
 
                                 for (let group of options) {
                                     if (group.group_id == group_id) {
-                                        option = group.twitter[subscribes[i].uid];
+                                        option = group.twitter[curr_s.uid];
                                         break;
                                     }
                                 }
                                 if (!option) throw `Twitter转发时出错，${group_id}这个组没有配置`;
                                 else post = opt_dict(option.post);
-                                
-                                if (option.bbq == true) {
-                                    bbq_group.push(group_id);
-                                    for (let group of summ) {
-                                        if (group.group_id == group_id) {
-                                            count = group.count;
-                                            break;
-                                        }
-                                    }
-                                }
 
                                 for (let tweet of tweet_list) {
                                     let status = checkStatus(tweet);
@@ -427,23 +452,20 @@ function checkTwiTimeline() {
                                         let addon = [];
                                         if (status != "retweet") {
                                             if (option.notice != undefined) addon.push(`${option.notice}`);
-                                            if (option.bbq == true) {
-                                                count++;
-                                                addon.push(`坑位号: ${count}`);
-                                            }
                                             url_list.push(url);
                                         }
                                         addon.push(url);
-                                        format(tweet).then(payload => {
+                                        const context = {group_id : group_id, message_type : "group"};
+                                        format(tweet, false, context).then(payload => {
                                             payload += `\n\n${addon.join("\n")}`
-                                            replyFunc({group_id : group_id, message_type : "group"}, payload);
+                                            replyFunc(context, payload);
                                         }).catch(err => console.error(err));
                                     }
                                 }
                             }
 
                             //不好办啊
-                            setTimeout(updateTwitter, 500, bbq_group, url_list, tweet_list, subscribes[i]);
+                            setTimeout(updateTwitter, 500, tweet_list, curr_s);
                         }
                     } catch(err) {
                         console.error(err, '\n', subscribes[i]);
@@ -454,12 +476,13 @@ function checkTwiTimeline() {
                 }, (check_interval-subscribes.length*1000)/subscribes.length);
             }
         });
-    }, check_interval)
+    }
+    setTimeout(refreshTimeline, 5000);
 
     function checkStatus(tweet) {
         let status = "";
         if ("retweeted_status" in tweet || "retweeted_status_id_str" in tweet || /^RT @/.test(tweet.full_text)) status = "retweet";
-        else if ("in_reply_to_status_id" in tweet && tweet.in_reply_to_status_id != null) status = "reply";
+        else if ("in_reply_to_status_id_str" in tweet && tweet.in_reply_to_status_id_str != null) status = "reply";
         else if ("media" in  tweet.entities && tweet.entities.media[0].type == "photo") status = "pic";
         else status = "origin"
 
@@ -477,34 +500,19 @@ function checkTwiTimeline() {
         return false;
     }
 
-    function updateTwitter(bbq_group, url_list, tweet_list, subscribe) {
+    function updateTwitter(tweet_list, subscribe) {
         mongodb(DB_PATH, {useUnifiedTopology: true}).connect().then(async mongo => {
             const twitter_db = mongo.db('bot').collection('twitter');
             await twitter_db.updateOne(
                 {_id : subscribe._id},
                 {$set : {tweet_id : tweet_list[0].id_str, name : tweet_list[0].user.name}})
                 .then(result => {
-                    if (result.result.nModified < 1) {
+                    if (result.result.ok != 1 && result.result.nModified < 1) {
                         console.error(tweet_list[0].id_str, subscribe.tweet_id, tweet_list[0].user.name,
                              result, "\n twitter_db update error during checkTwitter");
                     }
                 })
                 .catch(err => console.error(err + "\n twitter_db update error during checkTwitter"));
-
-            if (url_list.length > 0) {
-                const twe_sum = mongo.db('bot').collection('twe_sum');
-                for (let group_id of bbq_group) {
-                    await twe_sum.updateOne(
-                        {group_id : group_id}, 
-                        {$inc : {count : url_list.length}, $push : {list : {$each : url_list}}})
-                        .then(result => {
-                            if (result.result.nModified < 1) {
-                                console.error("\n twe_sum can't update during checkTwitter, group_id=" + group_id + url_list);
-                            }
-                        })
-                        .catch(err => console.error(err + "\n twe_sum update error during checkTwitter"));
-                }
-            }
             mongo.close();
         });
     }
@@ -567,110 +575,121 @@ function clearSubs(context, group_id) {
 /**
  * 整理tweet_obj
  * @param {object} tweet Tweet object
- * @param {string} from_user Twitter用户名
+ * @param {string} end_point 是否停止进一步挖掘
+ * @param {object} end_point 用于发送的context
  * @returns Promise  排列完成的Tweet String
  */
-async function format(tweet, end_point=false) {
+async function format(tweet, end_point = false, context = false) {
+    if (!tweet) return "Twitter转发时错误";
     let payload = [];
     let text = "";
     if('full_text' in tweet) text = tweet.full_text;
     else text = tweet.text;
-    if ("retweeted_status" in tweet) {
-        let rt_status = await format(tweet.retweeted_status)
-        payload.push(`来自${tweet.user.name}的Twitter\n转推了`, rt_status);
-        return payload.join("\n");
-    }
-    let pics = "";
-    let src = "";
-    if ("extended_entities" in tweet) {
-        for (entity in tweet.extended_entities) {
-            if (entity == "media") {
-                let media = tweet.extended_entities.media;
-                for (let i = 0; i < media.length; i++) {
-                    text = text.replace(media[i].url, "");
-                    if (media[i].type == "photo") {
-                        src = [media[i].media_url_https.substring(0, media[i].media_url_https.length-4),
-                            `?format=${media[i].media_url_https.substring(media[i].media_url_https.length-3, media[i].media_url_https.length)}&name=4096x4096`].join("");
-                        pics += await sizeCheck(src) ? `[CQ:image,cache=0,file=${src}]` : `[CQ:image,cache=0,file=${media[i].media_url_https}] 注：这不是原图`;
-                    }
-                    else if (media[i].type == "animated_gif") {
-                        try {
-                            await exec(`ffmpeg -i ${media[i].video_info.variants[0].url} -loop 0 -y ${__dirname}/temp.gif`)
-                                .then(async ({stdout, stderr}) => {
-                                    if (stdout.length == 0) {
-                                        if (fs.statSync(`${__dirname}/temp.gif`).size < MAX_SIZE) {
-                                            let gif = fs.readFileSync(`${__dirname}/temp.gif`);
-                                            let base64gif = Buffer.from(gif, 'binary').toString('base64');
-                                            pics += `[CQ:image,file=base64://${base64gif}]`;
+    text = text.replace(/&amp;/g, "&").replace(/&#91;/g, "[").replace(/&#93;/g, "]").replace(/&lt;/g, "<").replace(/&gt;/g, ">");
+
+    try {
+        if ("retweeted_status" in tweet) {
+            let rt_status = await format(tweet.retweeted_status, true);
+            payload.push(`来自${tweet.user.name}的Twitter\n转推了`, rt_status);
+            return payload.join("\n");
+        }
+        let pics = "";
+        let src = "";
+        if ("extended_entities" in tweet) {
+            for (entity in tweet.extended_entities) {
+                if (entity == "media") {
+                    let media = tweet.extended_entities.media;
+                    for (let i = 0; i < media.length; i++) {
+                        text = text.replace(media[i].url, "");
+                        if (media[i].type == "photo") {
+                            src = [media[i].media_url_https.substring(0, media[i].media_url_https.length-4),
+                                `?format=${media[i].media_url_https.substring(media[i].media_url_https.length-3, media[i].media_url_https.length)}&name=4096x4096`].join("");
+                            pics += `[CQ:image,cache=0,file=${src}]`;
+                        }
+                        else if (media[i].type == "animated_gif") {
+                            try {
+                                await exec(`ffmpeg -i ${media[i].video_info.variants[0].url} -loop 0 -y ${__dirname}/temp.gif`)
+                                    .then(async ({stdout, stderr}) => {
+                                        if (stdout.length == 0) {
+                                            if (fs.statSync(`${__dirname}/temp.gif`).size < MAX_SIZE) {
+                                                let gif = fs.readFileSync(`${__dirname}/temp.gif`);
+                                                let base64gif = Buffer.from(gif, 'binary').toString('base64');
+                                                pics += `[CQ:image,file=base64://${base64gif}]`;
+                                            }
+                                            else pics += `这是一张动图 [CQ:image,cache=0,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
                                         }
-                                        else pics += `这是一张动图 [CQ:image,cache=0,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
-                                    }
-                                })
-                        } catch(err) {
-                            console.error(err);
-                            pics += `这是一张动图 [CQ:image,cache=0,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
+                                    })
+                            } catch(err) {
+                                console.error(err);
+                                pics += `这是一张动图 [CQ:image,cache=0,file=${media[i].media_url_https}]` + `动起来看这里${media[i].video_info.variants[0].url}`;
+                            }
                         }
-                    }
-                    else if (media[i].type == "video") {
-                        let mp4obj = [];
-                        for (let j = 0; j < media[i].video_info.variants.length; j++) {
-                            if (media[i].video_info.variants[j].content_type == "video/mp4") mp4obj.push(media[i].video_info.variants[j]);
+                        else if (media[i].type == "video") {
+                            let mp4obj = [];
+                            for (let j = 0; j < media[i].video_info.variants.length; j++) {
+                                if (media[i].video_info.variants[j].content_type == "video/mp4") mp4obj.push(media[i].video_info.variants[j]);
+                            }
+                            mp4obj.sort((a, b) => {return b.bitrate - a.bitrate;});
+                            payload.push(`[CQ:image,cache=0,file=${media[i].media_url_https}]`);
+                            if (context) {
+                                replyFunc(context, `[CQ:video,file=${mp4obj[0].url}]`);
+                            }
+                            else payload.push(`视频地址: ${mp4obj[0].url}`);
                         }
-                        mp4obj.sort((a, b) => {return b.bitrate - a.bitrate;});
-                        payload.push(`[CQ:image,cache=0,file=${media[i].media_url_https}]`,
-                                     `视频地址: ${mp4obj[0].url}`);
                     }
                 }
             }
+            if (pics != "") payload.push(pics);
         }
-        if (pics != "") payload.push(pics);
-    }
-    if ("is_quote_status" in tweet && tweet.is_quote_status == true) {
-        let quote_tweet = await getSingleTweet(tweet.quoted_status_id_str);
-        payload.push("引用了", await format(quote_tweet));
-        text = text.replace(tweet.quoted_status_permalink.url, "");
-    }
-    if ("in_reply_to_status_id" in tweet && tweet.in_reply_to_status_id != null && !end_point) {
-        let reply_tweet = await getSingleTweet(tweet.in_reply_to_status_id_str);
-        payload.push("回复了", await format(reply_tweet, true));
-    }
-    if ("card" in tweet) {
-        // payload.push(tweet.binding_values.title.string_value, urlExpand(card.url));
-        if (/poll\dchoice/.test(tweet.card.name)) {
-            if ("image_large" in tweet.card.binding_values) {
-                payload.push(`[CQ:image,cache=0,file=${tweet.card.binding_values.image_large.url}]`);
-            }
-
-            let end_time = new Intl.DateTimeFormat('zh-Hans-CN', {month: 'short', day: '2-digit', hour: '2-digit', minute: '2-digit', timeZone: 'Asia/Shanghai'})
-                .format(new Date(tweet.card.binding_values.end_datetime_utc.string_value))
-            payload.push("", tweet.card.binding_values.counts_are_final.boolean_value === true ? "投票已结束" 
-                : `正在投票,结束时间${end_time}`);
-            let nchoice = parseInt(/\d/.exec(tweet.card.name)[0]);
-            let count = "";
-            let lable = "";
-            for (i = 1; i < nchoice + 1; i++) {
-                lable = tweet.card.binding_values[`choice${i}_label`].string_value;
-                count = tweet.card.binding_values[`choice${i}_count`].string_value;
-                payload.push(`${lable}:  ${count}`);
-            }
+        if (!end_point && "is_quote_status" in tweet && tweet.is_quote_status == true) {
+            let quote_tweet = await getSingleTweet(tweet.quoted_status_id_str);
+            payload.push("引用了", await format(quote_tweet, true));
+            text = text.replace(tweet.quoted_status_permalink.url, "");
         }
-        else if (/summary/.test(tweet.card.name)) {
-            if ("photo_image_full_size_original" in tweet.card.binding_values) {
-                if (sizeCheck(tweet.card.binding_values.photo_image_full_size_original.image_value.url)) {
+        if ("in_reply_to_status_id" in tweet && tweet.in_reply_to_status_id != null && !end_point) {
+            let reply_tweet = await getSingleTweet(tweet.in_reply_to_status_id_str);
+            payload.push("回复了", await format(reply_tweet, true));
+        }
+        if ("card" in tweet) {
+            if (/poll\dchoice/.test(tweet.card.name)) {
+                if ("image_large" in tweet.card.binding_values) {
+                    payload.push(`[CQ:image,cache=0,file=${tweet.card.binding_values.image_large.url}]`);
+                }
+    
+                let end_time = new Intl.DateTimeFormat('zh-Hans-CN', {month: 'short', day: '2-digit', hour: '2-digit', minute: '2-digit', timeZone: 'Asia/Shanghai'})
+                    .format(new Date(tweet.card.binding_values.end_datetime_utc.string_value))
+                payload.push("", tweet.card.binding_values.counts_are_final.boolean_value === true ? "投票已结束" 
+                    : `正在投票,结束时间${end_time}`);
+                let nchoice = parseInt(/\d/.exec(tweet.card.name)[0]);
+                let count = "";
+                let lable = "";
+                for (i = 1; i < nchoice + 1; i++) {
+                    lable = tweet.card.binding_values[`choice${i}_label`].string_value;
+                    count = tweet.card.binding_values[`choice${i}_count`].string_value;
+                    payload.push(`${lable}:  ${count}`);
+                }
+            }
+            else if (/summary/.test(tweet.card.name)) {
+                if ("photo_image_full_size_original" in tweet.card.binding_values) {
                     payload.push(`[CQ:image,cache=0,file=${tweet.card.binding_values.photo_image_full_size_original.image_value.url}]`);
                 }
-                else payload.push(`[CQ:image,cache=0,file=${tweet.card.binding_values.photo_image_full_size_large.image_value.url}]`);
+                if ("title" in tweet.card.binding_values) payload.push(tweet.card.binding_values.title.string_value)
+                if ("description" in tweet.card.binding_values) payload.push(tweet.card.binding_values.description.string_value);
             }
-            payload.push(tweet.card.binding_values.title.string_value, tweet.card.binding_values.description.string_value);
         }
-    }
-    if ("urls" in tweet.entities && tweet.entities.urls.length > 0) {
-        for (let i = 0; i <  tweet.entities.urls.length; i++) {
-            text = text.replace(tweet.entities.urls[i].url, tweet.entities.urls[i].expanded_url);
+        if ("urls" in tweet.entities && tweet.entities.urls.length > 0) {
+            for (let i = 0; i <  tweet.entities.urls.length; i++) {
+                text = text.replace(tweet.entities.urls[i].url, tweet.entities.urls[i].expanded_url);
+            }
         }
+        payload.unshift(`${tweet.user.name}的Twitter`, text);
+        return payload.join("\n");
     }
-    payload.unshift(`${tweet.user.name}的Twitter`, text);
-    return payload.join("\n");
+    catch (err) {
+        console.error(err);
+        payload.push(`${tweet.user.name}的Twitter`, text);
+        return payload.join("\n");
+    }
 }
 
 /**
@@ -696,11 +715,19 @@ function rtTimeline(context, name, num) {
         if (!user) replyFunc(context, "没这人");
         else if (user.protected == true) replyFunc(context, "这人的Twitter受保护");
         else {
-            getUserTimeline(user.id_str, 20).then(async timeline => {
-                if (timeline.length-1 < num) timeline = await getUserTimeline(user.id_str, 50);
-                if (timeline.length-1 < num) timeline = await getUserTimeline(user.id_str, 1, true, false);
-                format(timeline[num]).then(tweet_string => {
-                    let payload = [tweet_string, `https://twitter.com/${user.screen_name}/status/${timeline[num].id_str}`].join('\r\n\r\n');
+            getUserTimeline(user.id_str, 10).then(async timeline => {
+                let tweets = [];
+                for (let tweet of timeline) {
+                    if (!"retweeted_status_id_str" in tweet
+                    || !/^RT @/.test(tweet.full_text)) {
+                        tweets.push(tweet);
+                    }
+                }
+                if (tweets.length < num) tweets = timeline;
+                let choose_one = tweets[num];
+                choose_one.user = {name : user.name};
+                format(choose_one).then(tweet_string => {
+                    let payload = [tweet_string, `https://twitter.com/${user.screen_name}/status/${choose_one.id_str}`].join('\n\n');
                     replyFunc(context, payload);
                 }).catch(err => console.error(err));
             });
@@ -710,7 +737,7 @@ function rtTimeline(context, name, num) {
 
 function rtSingleTweet(tweet_id_str, context) {
     getSingleTweet(tweet_id_str).then(tweet => {
-        format(tweet).then(tweet_string => replyFunc(context, tweet_string))
+        format(tweet, false, context).then(tweet_string => replyFunc(context, tweet_string));
     });
 }